Small code cleanup
